Keep the template-to-fork control legal under elpscheck and -race
The control committed with the previous change was itself broken in two
ways, both of which the gate sweep caught after I had pushed on the
strength of the first three arms. Pushing before the sweep finished was the
mistake; the findings are real.

1. DATA RACE, of the control's own making. Every fork the broken walker
makes shares one map with the template, so the concurrent arm's two forks
writing that map in parallel is a genuine race -- reported against my
control, not against elps. It needs exactly one fork to demonstrate the
property, so it now supplies one transaction.

2. OWNERSHIP VIOLATION under -tags elpscheck: "LVal used by two Runtimes".
Both constructions were tried. Handing the fork the TEMPLATE's own *LVal
panics, which is correct and is not the defect being modelled. Binding a
FORK-OWNED header over the template's *MapData -- two headers over one
payload, the #576/#585 shape exactly -- panics too, because the values
inside the shared map are still the template's.

So under elpscheck, elps ALREADY refuses this class, and the control lives
in a `//go:build !elpscheck` file that says so. That bounds what the new
property adds, and the bound is worth stating: ownership checking is opt-in
and build-tag-gated, exactly like RuntimeBound, so in the ordinary build an
embedder ships, nothing else asserts that a write through the template
cannot reach a live fork.

Verified: green under both tag sets, green under -race, and still
load-bearing -- removing the property from the sweep turns the control red,
with go vet clean on the weakened tree.
